get a functional pushdown impl
